+++ piecrust-uplink/CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `ContractError::to_parts` and `ContractError::from_parts` functions [#301]
+- Add `fn_name` and `fn_arg` as an argument to `call_raw` and `call_raw_with_limit` [#301]
+
+### Changed
+
+- Rename `ContractError::Other` to `ContractError::Unknown` [#301]
+- Rename `ContractError::OutOfGas` to `ContractError::OutOfPoints` [#301]
+- Change `Display` for `ContractError` to display messages [#301]
+- Change `ContractError` variants to be CamelCase [#301]
+
+### Removed
+
+- Remove `ContractError::from_code` function [#301]
+- Remove `raw_call` as an argument of `call_raw` and `call_raw_with_limit` [#301]
+- Remove `RawCall` and `RawResult` [#301]
+
 ## [0.8.0] - 2023-10-11
 
 ### Added
@@ -120,6 +138,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - First `piecrust-uplink` release
 
 <!-- ISSUES -->
+[#301]: https://github.com/dusk-network/piecrust/issues/301
 [#271]: https://github.com/dusk-network/piecrust/issues/271
 [#268]: https://github.com/dusk-network/piecrust/issues/268
 [#243]: https://github.com/dusk-network/piecrust/issues/243